refactor: remove the home NavLink and add new components: MainInterface and JoinAShoppingList
