+++ src/components/JoinAShoppingList.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const JoinAShoppingList = () => {
+  return <div>JOIN A SHOPPING LIST</div>;
+};
+
+export default JoinAShoppingList;
